+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/SoraMediaChannel.kt
@@ -56,8 +56,8 @@ class SoraMediaChannel @JvmOverloads constructor(
         private var listener:                Listener?,
         private val clientId:                String?              = null,
         private val signalingNotifyMetadata: Any?                 = null,
-        private val peerConnectionOption:    PeerConnectionOption = PeerConnectionOption()
-        ) {
+        private val peerConnectionOption:    PeerConnectionOption = PeerConnectionOption()) {
+
     companion object {
         private val TAG = SoraMediaChannel::class.simpleName
 
@@ -411,7 +411,7 @@ class SoraMediaChannel @JvmOverloads constructor(
     private var timer: Timer? = null
 
     private fun startTimer() {
-       stopTimer()
+        stopTimer()
         timer = Timer()
         timer!!.schedule(object: TimerTask() {
             override fun run() {